requirements.txt
youtube_transcript_api
streamlit
transformers
torch
python-dotenv
pathlib
requests